builder restructure suggestion
